text-freetype2: Fix compiler warnings
CMake build framework 3.0 introduces more strict typing checks, which
require additional code fixes to silence.

* Fixes warnings about implicit casts
* Fixes warnings about shadowing existing variables

+++ plugins/text-freetype2/find-font.h
@@ -16,7 +16,7 @@ struct font_path_info {
 
 	bool is_bitmap;
 	uint32_t num_sizes;
-	int *sizes;
+	long *sizes;
 
 	bool bold;
 	bool italic;
